Factor out integration-test machinery.
We use some crazy nose magic here. Putting in it sphinx_js.testing makes nose puke when it finds a setup() in sphinx_js/__init__.py, so we shove it in tests/ and let the magic do its thing.
